Move self_and_owners to group module
