+++ gaphor/core/modeling/__init__.py
@@ -12,7 +12,7 @@ from gaphor.core.modeling.diagram import (
     DrawContext,
     UpdateContext,
 )
-from gaphor.core.modeling.element import Element, self_and_owners
+from gaphor.core.modeling.element import Element
 from gaphor.core.modeling.elementfactory import ElementFactory
 from gaphor.core.modeling.event import *
 from gaphor.core.modeling.presentation import Presentation
